+++ js/conversation.js
@@ -2,8 +2,9 @@ import { ConversationInputForm } from "./formConversationInput.js";
 import { FileInputForm } from "./formAddParticipant.js";
 import { ConversationEntrySheet } from "./conversationEntrySheet.js";
 import { getActorDataFromDragEvent, updateConversationControls, updateConversationLayout } from "./helpers.js";
+import { socket } from "./init.js";
 
-class ConversationHud {
+export class ConversationHud {
   // Function that initializes the class data
   init() {
     // Initialize variables
@@ -510,82 +511,3 @@ class ConversationHud {
     socket.executeForEveryone("updateActiveConversation", parsedData);
   }
 }
-
-async function checkConversationDataAvailability(users) {
-  // let userId = users[0].id;
-  for (let i = 0; i < users.length; i++) {
-    const result = await socket.executeAsUser("getActiveConversation", users[i].id);
-
-    // Check to see if we have a result and it is an active conversation
-    if (result) {
-      if (result.conversationIsActive) {
-        // We found an active conversation
-        const userId = users[i].id;
-        return { result, userId };
-      }
-    }
-  }
-}
-
-let socket;
-Hooks.once("socketlib.ready", () => {
-  socket = socketlib.registerModule("conversation-hud");
-});
-
-Hooks.on("init", () => {
-  game.ConversationHud = new ConversationHud();
-  game.ConversationHud.init();
-});
-
-Hooks.on("ready", async () => {
-  if (game.user.isGM) {
-    // Get a list of the connected users
-    const regularUsers = game.users.filter((item) => item.active && item.id !== game.user.id && !item.isGM);
-    const gmUsers = game.users.filter((item) => item.active && item.id !== game.user.id && item.isGM);
-
-    // Prioritize getting data from the GM users before trying to get data from the regular users
-    let conversationData;
-    if (gmUsers.length > 0) {
-      conversationData = await checkConversationDataAvailability(gmUsers);
-    }
-
-    // If we found no active conversation in the gm users, check to see if the regular users have one
-    if (!conversationData) {
-      if (regularUsers.length > 0) {
-        conversationData = await checkConversationDataAvailability(regularUsers);
-      }
-    }
-
-    if (conversationData) {
-      const visibility = await socket.executeAsUser("getActiveConversationVisibility", conversationData.userId);
-      game.ConversationHud.renderConversation(conversationData.result.activeConversation, visibility);
-    }
-  } else {
-    try {
-      // Get conversation data from a GM
-      const result = await socket.executeAsGM("getActiveConversation");
-
-      // If there is an active conversation, render it
-      if (result.conversationIsActive) {
-        const visibility = await socket.executeAsGM("getActiveConversationVisibility");
-        game.ConversationHud.renderConversation(result.activeConversation, visibility);
-      }
-    } catch (error) {
-      if (error.name === "SocketlibNoGMConnectedError") {
-        // If no GM is connected, try to get the data from another user
-        const users = game.users.filter((item) => item.active && item.id !== game.user.id);
-
-        if (users.length > 0) {
-          const { result, userId } = await checkConversationDataAvailability(users);
-
-          if (result) {
-            const visibility = await socket.executeAsUser("getActiveConversationVisibility", userId);
-            game.ConversationHud.renderConversation(result.activeConversation, visibility);
-          }
-        }
-      } else {
-        console.error(error);
-      }
-    }
-  }
-});
