+++ module.json
@@ -37,7 +37,7 @@
       }
     ]
   },
-  "esmodules": ["js/controls.js", "js/conversation.js", "js/entityLinkHandler.js", "js/hooks.js"],
+  "esmodules": ["js/init.js", "js/controls.js", "js/hooks.js"],
   "styles": ["css/main.css"],
   "languages": [
     {
